# .docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/compliance-organization/tenants": {
      "post": {
        "operationId": "ComplianceOrganizationController_upsertTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant upserted; includes default branch id and optional eTIMS row",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantUpsertResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create or update a compliance tenant (new tenants get a default Headquarters branch). Send `kraPin` (+ Test/Live) to also upsert the eTIMS shell on that branch.",
        "tags": [
          "Compliance organization"
        ]
      }
    },
    "/compliance-organization/tenants/by-sync2books/{sync2booksCompanyId}": {
      "get": {
        "operationId": "ComplianceOrganizationController_getTenantBySync2books",
        "parameters": [
          {
            "name": "sync2booksCompanyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant or 404"
          }
        },
        "summary": "Get tenant by Sync2Books company id",
        "tags": [
          "Compliance organization"
        ]
      }
    },
    "/compliance-organization/tenants/{tenantId}": {
      "get": {
        "operationId": "ComplianceOrganizationController_getTenantById",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get tenant by internal id",
        "tags": [
          "Compliance organization"
        ]
      }
    },
    "/compliance-organization/tenants/{tenantId}/branches": {
      "post": {
        "operationId": "ComplianceOrganizationController_upsertBranch",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create or update a branch under a tenant",
        "tags": [
          "Compliance organization"
        ]
      },
      "get": {
        "operationId": "ComplianceOrganizationController_listBranches",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List branches for a tenant",
        "tags": [
          "Compliance organization"
        ]
      }
    },
    "/compliance-organization/branches/{branchId}/etims-connection": {
      "get": {
        "operationId": "ComplianceOrganizationController_getEtimsConnection",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "eTIMS connection or 404"
          }
        },
        "summary": "Get the eTIMS connection (kraPin, environment, status, dvcSrlNo) for a branch. Used by the main API to reconcile locally-denormalized fields (e.g. backfill).",
        "tags": [
          "Compliance organization"
        ]
      },
      "put": {
        "operationId": "ComplianceOrganizationController_upsertEtimsConnection",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertEtimsConnectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Create or update eTIMS connection shell (kraPin, environment, etc.). OSCU deviceId/cmcKey are set only after initialize, not from this body.",
        "tags": [
          "Compliance organization"
        ]
      }
    },
    "/compliance-organization/branches/{branchId}/etims-connection/initialize": {
      "post": {
        "operationId": "ComplianceOrganizationController_initializeEtimsConnection",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection updated after initialize"
          }
        },
        "summary": "Call OSCU initialize using stored connection (kraPin, dvcSrlNo) and branch kraBhfId; persist cmcKey + dvcId from response. No body — branch id in path only.",
        "tags": [
          "Compliance organization"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sync2Books Compliance API",
    "description": "Unified compliance abstraction layer for eTIMS (OSCU).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpsertTenantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "nullable": true,
            "description": "Internal Compliance tenant id. Send when updating a tenant created without a Sync2Books company id."
          },
          "sync2booksCompanyId": {
            "type": "object",
            "nullable": true,
            "description": "Sync2Books company id when integrating with the main Sync2Books API; omit for compliance-dashboard-only tenants."
          },
          "displayName": {
            "type": "object",
            "nullable": true
          },
          "kraPin": {
            "type": "object",
            "nullable": true,
            "description": "KRA PIN / TIN (DigiTax “PIN No.”). When set, creates or updates the eTIMS connection shell on the default Headquarters branch.",
            "example": "A123456789B"
          },
          "environment": {
            "type": "string",
            "enum": [
              "SANDBOX",
              "PRODUCTION"
            ],
            "description": "SANDBOX = Test, PRODUCTION = Live. Takes precedence over `isLiveBusiness`."
          },
          "isLiveBusiness": {
            "type": "boolean",
            "description": "false = Test (SANDBOX), true = Live (PRODUCTION). Ignored when `environment` is set."
          }
        }
      },
      "TenantUpsertResponseDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "object"
          },
          "defaultBranchId": {
            "type": "string",
            "description": "Internal id of the default branch (Headquarters), used for eTIMS routes."
          },
          "etimsConnection": {
            "type": "object",
            "nullable": true,
            "description": "Present when `kraPin` was sent: eTIMS connection shell on the default branch."
          }
        },
        "required": [
          "tenant",
          "defaultBranchId"
        ]
      },
      "UpsertBranchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "nullable": true,
            "description": "Internal branch id. Send when updating a branch that has no Sync2Books branch id."
          },
          "sync2booksBranchId": {
            "type": "object",
            "nullable": true,
            "description": "Sync2Books branch / location id when linked to the main product; omit for dashboard-only branches."
          },
          "displayName": {
            "type": "object",
            "nullable": true
          },
          "kraBhfId": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "UpsertEtimsConnectionDto": {
        "type": "object",
        "properties": {
          "kraPin": {
            "type": "string"
          },
          "dvcSrlNo": {
            "type": "object",
            "nullable": true,
            "description": "Device serial sent in initialize request body (`dvcSrlNo`). Not the same as OSCU `dvcId` (stored after initialize)."
          },
          "environment": {
            "type": "string",
            "enum": [
              "SANDBOX",
              "PRODUCTION"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "ERROR"
            ]
          },
          "sync2booksConnectionId": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "kraPin",
          "environment"
        ]
      },
...
    }
  }
}